+++ src/awsS3/signed.nim
@@ -2,13 +2,16 @@
 
 
 import
-  std/httpclient,
-  std/httpcore,
-  std/json,
-  std/mimetypes,
-  std/os,
-  std/strutils,
-  std/tables
+  std/[
+    httpclient,
+    httpcore,
+    json,
+    mimetypes,
+    os,
+    strutils,
+    tables,
+    times
+  ]
 
 import
   awsSigV4,
@@ -23,6 +26,10 @@ type
 const
   mimetypeDB = mimes.toTable
 
+const
+  #dateISO8601* = initTimeFormat "yyyyMMdd"
+  basicISO8601* = initTimeFormat "yyyyMMdd\'T\'HHmmss\'Z\'"
+
 
 proc s3SignedUrl*(
     credsAccessKey, credsSecretKey, credsRegion: string,
@@ -31,7 +38,8 @@ proc s3SignedUrl*(
     contentDisposition: S3ContentDisposition = CDTignore, contentDispositionName = "",
     setContentType = true,
     fileExt = "", customQuery = "", copyObject = "", expireInSec = "65",
-    accessToken = ""
+    accessToken = "",
+    makeDateTime = ""
   ): string =
   ## Generate a S3 signed URL.
   ##
@@ -43,91 +51,81 @@ proc s3SignedUrl*(
   ## fileExt => ".jpg", ".ifc"
 
   let
-    accessKey = credsAccessKey
-    secretKey = credsSecretKey
-    tokenKey  = accessToken
+    url = "https://" & bucketHost & "/" & key
+    region = credsRegion
+    service = "s3"
+    payload = ""
+    digest = SHA256
 
-    url       = "https://" & bucketHost & "/" & key
-    region    = credsRegion
-    service   = "s3"
+  # datetime:
+  #
+  # Why this? In a complex threaded system Valgrind kept bugging over the
+  # times library and not bein able to free the memory. The original
+  # makeDateTime() comes from the library awsSigV4, and even with destroy and
+  # defer nothing helped on Valgrind.
+  #
+  # You might never experience this, but if you do, the fix is to create the
+  # datetime string outside the procedure within a scoped block and just pass
+  # the string.
+  var datetime: string = makeDateTime
+  if datetime.len == 0:
+    datetime = getTime().utc.format(basicISO8601)
 
-    payload   = ""
-    digest    = SHA256
-    expireSec = expireInSec
-    datetime  = makeDateTime()
-    scope     = credentialScope(region=region, service=service, date=datetime)
+  let scope = credentialScope(region=region, service=service, date=datetime)
 
-  var
-    headers = newHttpHeaders(@[
-      ("Host", bucketHost)
-    ])
-
-  # Attach copyObject to headers
-  if copyObject != "":
+  var headers = newHttpHeaders()
+  headers.add("Host", bucketHost)
+  if copyObject.len > 0:
     headers.add("x-amz-copy-source", copyObject)
 
-  var
-    query = %*{
-              "X-Amz-Algorithm": $SHA256,
-              "X-Amz-Credential": accessKey & "/" & scope,
-              "X-Amz-Date": datetime,
-              "X-Amz-Expires": expireSec,
-              # "X-Amz-SignedHeaders": "host"
-            }
+  # Create the initial JSON query with known fields using %*
+  var query = %* {
+    "X-Amz-Algorithm": $SHA256,
+    "X-Amz-Credential": credsAccessKey & "/" & scope,
+    "X-Amz-Date": datetime,
+    "X-Amz-Expires": expireInSec
+  }
 
-  if tokenKey != "":
-    query["X-Amz-Security-Token"] = newJString(tokenKey)
+  if accessToken.len > 0:
+    query["X-Amz-Security-Token"] = %* accessToken
 
+  if contentDisposition != CDTignore or contentDispositionName.len > 0:
+    let dispType = case contentDisposition
+      of CDTinline: "inline;"
+      of CDTattachment: "attachment;"
+      else: ""
 
-  if contentDisposition != CDTignore or contentDispositionName != "":
-    let dispType =
-        case contentDisposition
-        of CDTinline:
-          "inline;"
-        of CDTattachment:
-          "attachment;"
-        else:
-          ""
-
-    let filename =
-        if contentDispositionName == "":
-          ""
-        elif dispType == "":
-          "filename=\"" & contentDispositionName & "\""
-        else:
-          " filename=\"" & contentDispositionName & "\""
-
-    query["response-content-disposition"] = newJString(dispType & filename)
-
+    if contentDispositionName.len > 0:
+      let filename = if dispType.len == 0:
+        "filename=\"" & contentDispositionName & "\""
+      else:
+        " filename=\"" & contentDispositionName & "\""
+      query["response-content-disposition"] = %* (dispType & filename)
 
   if setContentType:
-    let extension =
-        if fileExt != "":
-          fileExt[1..^1]
-        else:
-          splitFile(key).ext[1..^1]
+    let extension = if fileExt.len > 0: fileExt[1..^1]
+                    else: splitFile(key).ext[1..^1]
+    query["response-content-type"] = %* mimetypeDB.getOrDefault(extension, "binary/octet-stream")
 
-    query["response-content-type"] = newJString(mimetypeDB.getOrDefault(extension, "binary/octet-stream"))
+  if customQuery.len > 0:
+    for c in customQuery.split(","):
+      let q = c.split(":")
+      if q.len == 2:
+        query[q[0]] = %* q[1]
 
-
-  if customQuery != "":
-    for c in split(customQuery, ","):
-      let q = split(c, ":")
-      query[q[0]] = newJString(q[1])
-
-
-  # Add the signed headers to query
-  if copyObject != "":
-    query["X-Amz-SignedHeaders"] = newJString("host;x-amz-copy-source")
-  else:
-    query["X-Amz-SignedHeaders"] = newJString("host")
-
+  query["X-Amz-SignedHeaders"] = %* (if copyObject.len > 0: "host;x-amz-copy-source" else: "host")
 
   let
-    request   = canonicalRequest(httpMethod, url, query, headers, payload, digest = UnsignedPayload)
-    sts       = stringToSign(request, scope, date = datetime, digest = digest)
-    signature = calculateSignature(secret=secretKey, date = datetime, region = region,
-                                  service = service, tosign = sts, digest = digest)
+    request = canonicalRequest(httpMethod, url, query, headers, payload, digest = UnsignedPayload)
+    sts = stringToSign(request, scope, datetime, digest)
+    signature = calculateSignature(
+      secret = credsSecretKey,
+      date = datetime,
+      region = region,
+      service = service,
+      tosign = sts,
+      digest = digest
+    )
 
   result = url & "?" & request.split("\n")[2] & "&X-Amz-Signature=" & signature
 
@@ -140,7 +138,7 @@ proc s3SignedUrl*(awsCreds: AwsCreds, bucketHost, key: string,
     contentDisposition: S3ContentDisposition = CDTignore, contentDispositionName = "",
     setContentType = true, fileExt = "", customQuery = "", copyObject = "",
     expireInSec = "65"
-  ): string =
+  ): string {.deprecated.} =
 
   return s3SignedUrl(
       awsCreds.AWS_ACCESS_KEY_ID, awsCreds.AWS_SECRET_ACCESS_KEY, awsCreds.AWS_REGION,
@@ -161,7 +159,7 @@ proc s3Presigned*(accessKey, secretKey, region: string, bucketHost, key: string,
     httpMethod = HttpGet,
     contentDisposition: S3ContentDisposition = CDTattachment, contentDispositionName = "",
     setContentType = true, fileExt = "", expireInSec = "65", accessToken = ""
-  ): string =
+  ): string {.deprecated.} =
   ## Generates a S3 presigned url for sharing.
   ##
   ## contentDisposition => sets "Content-Disposition" type (inline/attachment)
@@ -180,7 +178,7 @@ proc s3Presigned*(accessKey, secretKey, region: string, bucketHost, key: string,
 
 proc s3Presigned*(creds: AwsCreds, bucketHost, key: string,
     contentDisposition: S3ContentDisposition = CDTattachment, contentDispositionName="",
-    setContentType=true, fileExt="", expireInSec="65"): string =
+    setContentType=true, fileExt="", expireInSec="65"): string {.deprecated.} =
 
   return s3Presigned(
       creds.AWS_ACCESS_KEY_ID, creds.AWS_SECRET_ACCESS_KEY, creds.AWS_REGION,